requirements.txt
asgiref==3.5.0
Django==4.0.3
psycopg2==2.9.3
sqlparse==0.4.2
Pillow==9.1.0
django-crispy-forms==1.14.0
